imooc-llmops-api/imooc-llmops-api-master/requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiolimiter==1.1.0
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.1
azure-core==1.30.2
backcall==0.2.0
backoff==2.2.1
bce-python-sdk==0.9.17
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
boto3==1.34.156
botocore==1.34.156
bs4==0.0.1
cachetools==5.4.0
catalogue==2.0.10
celery==5.3.6
certifi>=2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.18.1
cohere==5.6.2
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.0
cos-python-sdk-v5==1.9.31
crcmod==1.7
cryptography==42.0.8
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.6
decorator==5.1.1
deepdiff==7.0.1
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
duckduckgo_search==6.2.6
einops==0.8.0
email-validator==2.3.0
emoji==2.12.1
et-xmlfile==1.1.0
eventlet==0.37.0
exceptiongroup==1.2.0
executing==2.0.1
fastavro==1.9.5
fastjsonschema==2.19.1
filelock==3.15.4
filemagic==1.6
faiss-cpu==1.13.2
filetype==1.2.0
Flask==3.0.2
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec>=2024.6.0
future==1.0.0
gevent==23.9.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-health-checking==1.64.1
grpcio-tools==1.64.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.6
iniconfig==2.0.0
injector==0.21.0
intel-openmp==2021.4.0
iopath==0.1.10
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
kombu==5.4.2
langchain==0.2.16
langchain-community==0.2.10
langchain-core==0.2.41
langchain-experimental==0.0.63
langchain-huggingface==0.0.3
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langchain-weaviate==0.0.2
langchain-ollama==0.1.3
langcodes==3.4.0
langdetect==1.0.9
langgraph==0.2.8
langgraph-checkpoint==1.0.3
langsmith==0.1.129
language_data==1.2.0
lark==1.1.9
layoutparser==0.3.4
llvmlite==0.43.0
loguru==0.7.2
lxml==4.9.4
mailbox==0.4
Mako==1.3.2
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
nomic==3.1.2
numba==0.60.0
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.19.2
openai==1.37.0
opencv-python==4.10.0.84
openpyxl==3.1.5
ordered-set==4.1.0
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
phonenumbers==8.13.40
pickleshare==0.7.5
pillow==10.4.0
pillow_heif==0.18.0
pinecone-client==3.2.2
pipreqs==0.5.0
platformdirs==4.2.0
pluggy==1.4.0
portalocker==2.10.1
preshed==3.0.9
presidio_analyzer==2.2.355
presidio_anonymizer==2.2.355
primp==0.5.5
prompt-toolkit==3.0.43
protobuf==5.27.2
psutil==6.0.0
psycopg2==2.9.9
pure-eval==0.2.2
pyarrow==17.0.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.17.2
PyJWT==2.9.0
pynndescent==0.5.13
pyparsing==3.2.0
pypdf==4.3.0
pypdfium2==4.30.0
pyreadline3==3.5.4
pytest==8.1.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639>=2024.4.27
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.12
python-pptx==0.6.23
pytz==2025.1
dashscope==1.20.0
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
qianfan==0.3.17.1
rank-bm25==0.2.2
rapidfuzz==3.9.4
redis==5.0.8
referencing==0.34.0
regex>=2024.5.15
RestrictedPython==7.4
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.0
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
shellingham==1.5.4
simsimd==4.4.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
srsly==2.4.8
stack-data==0.6.3
sympy==1.13.0
tabulate==0.9.0
tbb==2021.13.0
tcvectordb==1.8.5
tenacity==8.3.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.9
tinycss2==1.2.1
tldextract==5.1.2
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.1
torchvision==0.19.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.41.2
typer==0.12.3
types-requests>=2.32.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2025.1
umap-learn==0.5.6
unstructured==0.15.0
unstructured-client==0.24.1
unstructured-inference==0.7.40
# urllib3
validators==0.34.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
weaviate-client==4.8.1
webencodings==0.5.1
Werkzeug==3.0.1
wikipedia==1.4.0
win32-setctime==1.1.0
wrapt==1.16.0
WTForms==3.1.2
XlsxWriter==3.2.0
xmltodict==0.13.0
yarg==0.1.9
yarl==1.9.4
zope.event==5.0
zope.interface==7.0.3
